+++ frontend/src/app/pages/registro-revisao/registro-revisao.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { Revisao } from '../../core/models/revisao.model';
 import { RevisaoService } from '../../core/services/revisao.service';
 import { ToastService } from '../../core/services/toast.service';
+import { ConfirmationDialogService } from '../../core/services/confirmationDialog.service';
 
 @Component({
   selector: 'app-registro-revisao',
@@ -17,7 +18,8 @@ export class RegistroRevisaoComponent {
   loading: boolean = false;
 
   constructor(private revisaoService: RevisaoService,
-    private toastService: ToastService
+    private toastService: ToastService,
+    private confirmationService: ConfirmationDialogService
   ) { }
 
   ngOnInit(): void {
@@ -49,18 +51,31 @@ export class RegistroRevisaoComponent {
   }
 
   removeRevisao(id: number): void {
-    this.loading = true;
-    this.revisaoService.delete(`/${id}`).subscribe(() => {
-      this.loading = false;
-      this.revisoes = this.revisoes.filter(r => r.idRevisao !== id);
-      this.onSearch();
-      this.closeModal();
-      this.toastService.showSuccess("Revisão removida com sucesso!")
-      this.loading = false;
-    });
+    this.confirmationService.confirm(
+      'Tem certeza?',
+      'Você deseja realmente excluir este item?'
+    ).subscribe(result => {
+      this.loading = true;
+      if (!result) return;
+      this.revisaoService.delete(`/${id}`).subscribe(() => {
+        this.loading = false;
+        this.revisoes = this.revisoes.filter(r => r.idRevisao !== id);
+        this.onSearch();
+        this.closeModal();
+        this.toastService.showSuccess("Revisão removida com sucesso!")
+        this.loading = false;
+      });
+    })
   }
 
   editRevisao(id: number): void {
-
+    this.confirmationService.confirm(
+      'Tem certeza?',
+      'Você deseja realmente excluir este item?'
+    ).subscribe(result => {
+      if (result) {
+        console.log("positivo")
+      }
+    });
   }
 }
